Use consistent naming, also check attributes outside polymer databindings.
