+++ src/collections.ts
@@ -22,7 +22,7 @@ registry.register(
       'dom-module-invalid-attrs',
       'style-into-template',
       'unbalanced-polymer-delimiters',
-      'unknown-set-attribute',
+      'set-unknown-attribute',
     ]));
 
 registry.register(new RuleCollection(
@@ -34,7 +34,7 @@ Will warn about use of deprecated Polymer 1.x features or brand new features in 
       'dom-module-invalid-attrs',
       'style-into-template',
       'unbalanced-polymer-delimiters',
-      'unknown-set-attribute',
+      'set-unknown-attribute',
     ]));
 
 registry.register(new RuleCollection(
@@ -45,5 +45,5 @@ For projects that are ready to start transitioning to Polymer 2.0 see polymer-2-
 `,
     [
       'unbalanced-polymer-delimiters',
-      'unknown-set-attribute',
+      'set-unknown-attribute',
     ]));
