+++ src/rules.ts
@@ -15,4 +15,4 @@
 import './html/dom-module-name-or-is';
 import './html/move-style-into-template';
 import './polymer/unbalanced-delimiters';
-import './polymer/bind-to-undeclared';
+import './polymer/set-unknown-attribute';
