+++ test/set-unknown-attribute/when-to-warn.html
@@ -64,3 +64,6 @@
     -->
   </template>
 </dom-module>
+
+<elem-one attr-one="{{fine because this isn't a databinding}}"></elem-one>
+<elem-one unknown="warns even outside databindings"></elem-one>
